pyproject.toml
[project]
name = "dsproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "altair>=6.0.0",
    "mysql-connector-python>=9.5.0",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "stats-can>=3.1.0",
    "streamlit>=1.53.0",
]
